+++ javascript/node/selenium-webdriver/CHANGES.md
@@ -3,6 +3,8 @@
 * Added the `selenium-webdriver/testing` package, which provides a basic
     framework for writing tests using Mocha. See
     `selenium-webdriver/example/google_search_test.js` for usage.
+* For Promises/A+ compatibility, backing out the change in 2.30.0 that ensured
+    rejections were always Error objects. Rejection reasons are now left as is.
 * FIXED: remote.SeleniumServer#stop now shuts down within the active control
     flow, allowing scripts to finish. Use #kill to shutdown immediately.